Add user_input method directly to run.rb
