+++ operations.rb
@@ -2,18 +2,6 @@ require_relative './store'
 
 class Operations
 
-  def user_input
-    puts 'Welcome to School Library App'
-    puts 'Please choose an option by entering a number:'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - Exit'
-  end
-
   def list_books
     store = Store.new
     books_arr = store.books_arr
